Made LED blinking using memory-mapped IO
Since this version doesn't have the correct alarm or badTry functions, I'm not 100% sure it works. It should, but if it doesn't you can just revert them back to digitalWrite(12, HIGH or LOW). The only things that need to be added now are the alarm and badTry functions. Other than that I think it is pretty much finished.